start more opcode

+++ gbaemu-common/src/mem/mod.rs
@@ -1,4 +1,4 @@
 pub trait Memory: std::fmt::Debug {
-    fn read(&self, address: u16) -> u16;
+    fn read(&self, address: u32) -> u32;
     fn write(&mut self, address: u16, value: u16);
 }
